Add files via upload
Project
The project aims to develop a predictive model for used car data that can help buyers, sellers, and dealers to estimate the fair value of a used car. The model will use historical data on used car sales, including factors such as the car's age, mileage, make and model, and condition, to predict the market value of a particular used car. Along with the price we calculate the profit based on the selling price, year of manufacture and original car price. Calculated with depreciation factor.